+++ code/JavaRecursionExamples/mutualrec/non_term/two_derived_explicit_decoration/TwoDerivedExplicitDecoration.java
@@ -1,4 +1,4 @@
-package mutualrec.non_term_2_derived;
+package mutualrec.non_term.two_derived_explicit_decoration;
 
 /*
 +package sandbox
@@ -14,15 +14,11 @@ package mutualrec.non_term_2_derived;
 
 [] > derived1
   base > @
-  [self v] > m
-    self.n self v > @
   [self v] > n
     self.m self v > @
 
 [] > derived2
   base > @
-  [self v] > m
-    self.n self v > @
   [self v] > n
     self.m self v > @
  */
@@ -40,10 +36,6 @@ class Base {
 }
 
 class Derived1 extends Base {
-    @Override
-    public void m(int v) {
-        this.n(v);
-    }
 
     @Override
     public void n(int v) {
@@ -52,10 +44,6 @@ class Derived1 extends Base {
 }
 
 class Derived2 extends Base {
-    @Override
-    public void m(int v) {
-        this.n(v);
-    }
 
     @Override
     public void n(int v) {
@@ -63,13 +51,12 @@ class Derived2 extends Base {
     }
 }
 
-public class NonTerm2Derived {
 
+public class TwoDerivedExplicitDecoration {
     public static void main(String[] args) {
         Derived1 derived1Instance = new Derived1();
         Derived2 derived2Instance = new Derived2();
         derived1Instance.m(12);
         derived2Instance.m(12);
     }
-
 }
